integrated song playing in lyricsDisplayPage and also integrated spotify web api libraries
