+++ app/build.gradle
@@ -28,22 +28,21 @@ android {
     }
 }
 
+repositories {
+    mavenCentral()
+    maven { url "https://jitpack.io" }
+}
+
 dependencies {
 
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation fileTree(include: ['*.jar'], dir: 'app')
     implementation("com.squareup.okhttp3:okhttp:4.9.0")
 
-
-    // your app dependencies
-    implementation project(':spotify-app-remote-release-0.7.2')
-
     // spotify api dependencies
     implementation project(':spotify-app-remote-release-0.7.2')
     implementation 'com.spotify.android:auth:1.1.0'
-    // your app dependencies
-    implementation "com.google.code.gson:gson:2.8.5"
-
+    implementation 'com.github.kaaes:spotify-web-api-android:0.4.1'
     implementation 'androidx.appcompat:appcompat:1.3.0'
     implementation 'com.google.android.material:material:1.4.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
