+++ app/src/main/java/com/example/spotifykaraokeee/LyricDisplayPage.java
@@ -4,18 +4,25 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 import android.text.method.ScrollingMovementMethod;
+import android.util.Log;
 import android.widget.TextView;
 
-import com.amazonaws.mobileconnectors.apigateway.ApiClientFactory;
-import com.spotifykaraoke.clientsdk.GetLyricsClient;
+
+import com.spotify.android.appremote.api.ConnectionParams;
+import com.spotify.android.appremote.api.Connector;
+import com.spotify.android.appremote.api.SpotifyAppRemote;
+import com.spotify.sdk.android.authentication.AuthenticationResponse;
 
 import org.jetbrains.annotations.NotNull;
-
-import java.io.BufferedInputStream;
 import java.io.IOException;
-import java.io.InputStream;
-import java.net.HttpURLConnection;
-import java.net.URL;
+import java.util.List;
+
+import kaaes.spotify.webapi.android.SpotifyApi;
+import kaaes.spotify.webapi.android.SpotifyCallback;
+import kaaes.spotify.webapi.android.SpotifyError;
+import kaaes.spotify.webapi.android.SpotifyService;
+import kaaes.spotify.webapi.android.models.Track;
+import kaaes.spotify.webapi.android.models.TracksPager;
 
 import okhttp3.Call;
 import okhttp3.Callback;
@@ -23,7 +30,17 @@ import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
 
+
 public class LyricDisplayPage extends AppCompatActivity {
+
+
+    // Set up Spotify Credentials
+    private static final int REQUEST_CODE = 1337;
+    //our clientID from our developer dashboard
+    private static final String CLIENT_ID = "b6cb228259164dabab54198bf1537a75";
+    //also had to set this as our redirect URI in our developer dashboard
+    private static final String REDIRECT_URI = "http://com.yourdomain.yourapp/callback";
+    private SpotifyAppRemote mSpotifyAppRemote;
 
     private final OkHttpClient client = new OkHttpClient();
     @Override
@@ -35,6 +52,31 @@ public class LyricDisplayPage extends AppCompatActivity {
         String artist = "Michael Jackson";
 
         displayLyrics(song_title, artist);
+        playSong();
+
+        /*
+        SpotifyApi api = new SpotifyApi();
+
+        api.setAccessToken();
+
+        SpotifyService spotify = api.getService();
+
+        spotify.searchTracks(song_title, new SpotifyCallback<TracksPager>() {
+            @Override
+            public void success(TracksPager tracksPager, retrofit.client.Response response) {
+               List<Track> songs = tracksPager.tracks.items;
+                //System.out.print(songs);
+            }
+
+            @Override
+            public void failure(SpotifyError spotifyError) {
+                Log.d("Track failure", spotifyError.toString());
+            }
+        });
+*/
+
+
+
 
     }
     //this method makes the http GET request to our api endpoint using OkHTTP library
@@ -69,6 +111,49 @@ public class LyricDisplayPage extends AppCompatActivity {
                 e.printStackTrace();
             }
         });
+    }
+
+    private void playSong(){
+
+        ConnectionParams connectionParams =
+                new ConnectionParams.Builder(CLIENT_ID)
+                        .setRedirectUri(REDIRECT_URI)
+                        .showAuthView(true)
+                        .build();
+
+        SpotifyAppRemote.connect(this, connectionParams, new Connector.ConnectionListener() {
+            @Override
+            public void onConnected(SpotifyAppRemote spotifyAppRemote) {
+                mSpotifyAppRemote = spotifyAppRemote;
+                Log.d("MainActivity", "Connected! Yay!");
+
+                // Now you can start interacting with App Remote
+                connected();
+
+            }
+            @Override
+            public void onFailure(Throwable error) {
+                // Something went wrong if you go here!
+                Log.e("MyActivity", error.getMessage(), error); // edited from throwable.getMessage()
+            }
+        });
+    }
+
+    // Connected functionality for remote play
+    private void connected() {
+
+        //plays default song from playlist in quick start guide
+        mSpotifyAppRemote.getPlayerApi().play("spotify:playlist:37i9dQZF1DX2sUQwD7tbmL");
+
+        //returns player state
+        mSpotifyAppRemote.getPlayerApi()
+                .subscribeToPlayerState()
+                .setEventCallback(playerState -> {
+                    final com.spotify.protocol.types.Track track = playerState.track;
+                    if (track != null) {
+                        Log.d("MainActively", track.name + "by + " + track.artist.name);
+                    }
+                });
     }
 }
