+++ app/src/main/java/com/example/spotifykaraokeee/ui/SearchSong/SearchSongFragment.java
@@ -1,6 +1,7 @@
 package com.example.spotifykaraokeee.ui.SearchSong;
 
 import android.os.Bundle;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -13,6 +14,15 @@ import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
 
 import com.example.spotifykaraokeee.R;
+
+import java.util.List;
+
+import kaaes.spotify.webapi.android.SpotifyApi;
+import kaaes.spotify.webapi.android.SpotifyCallback;
+import kaaes.spotify.webapi.android.SpotifyError;
+import kaaes.spotify.webapi.android.SpotifyService;
+import kaaes.spotify.webapi.android.models.Track;
+import kaaes.spotify.webapi.android.models.TracksPager;
 
 public class SearchSongFragment extends Fragment {
 
@@ -32,5 +42,28 @@ public class SearchSongFragment extends Fragment {
             }
         });
         return root;
+
+        /*
+        String song_title = "Love never felt so good";
+
+        SpotifyApi api = new SpotifyApi();
+
+        api.setAccessToken();
+
+        SpotifyService spotify = api.getService();
+
+        spotify.searchTracks(song_title, new SpotifyCallback<TracksPager>() {
+            @Override
+            public void success(TracksPager tracksPager, retrofit.client.Response response) {
+               List<Track> songs = tracksPager.tracks.items;
+                //System.out.print(songs);
+            }
+
+            @Override
+            public void failure(SpotifyError spotifyError) {
+                Log.d("Track failure", spotifyError.toString());
+            }
+        });
+        */
     }
 }
